dorm-power-alert/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
PySide6>=6.7.0
